app/health-server/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
groq
pillow
python-multipart
requests
